addCIModule OK
getCIMOdule(id) en deleteCIModule(id) werkend
Let op: id in model is Long en geen int. Ook Long aangepast in Television

+++ src/main/java/com/example/demo/repositories/TelevisionRepository.java
@@ -3,7 +3,7 @@ package com.example.demo.repositories;
 import com.example.demo.models.Television;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TelevisionRepository extends JpaRepository<Television, Integer> {
+public interface TelevisionRepository extends JpaRepository<Television, Long> {
 
     Iterable<Television> findAllByName(String name);
 //    Iterable<Television> findById(int id);
